`@remotion/whisper-web`: Fix Worker path

+++ packages/whisper-web/build-wasm.ts
@@ -93,12 +93,16 @@ if (!mainContent.includes('new Worker(new URL(')) {
 
 const workerContent =
 	content.replace(
-		'worker=new Worker(pthreadMainJs)',
+		'worker=new Worker(pthreadMainJs,workerOptions)',
 		`throw new Error('Already is in worker')`,
 	) +
 	'\n' +
 	'export default Module;' +
 	'\n';
+
+if (!workerContent.includes('Already is in worker')) {
+	throw new Error('Changes have not been made');
+}
 
 // Write the modified content directly to the destination
 fs.writeFileSync(path.join(__dirname, 'main.js'), mainContent, 'utf8');
